// emm/web_ui/src/api/auth/types.ts
export type Iso8601 = string;
// ...
export interface DockerRegistryConfig {
  department_id: string;
  is_enabled: boolean;
  pull_policy: "all" | "restricted";
  pull_user_ids: string[];
  push_user_ids: string[];
  created_at: Iso8601;
  updated_at?: Iso8601;
}

export interface DockerRegistryConfigPutRequest {
  pull_policy: "all" | "restricted";
  pull_user_ids?: string[];
  push_user_ids?: string[];
}

export interface DockerRegistryConfigPatchRequest
  extends Partial<DockerRegistryConfigPutRequest> {
  is_enabled?: boolean;
}

export interface DockerTokenResponse {
  token: string;
  access_token: string;
  expires_in: number;
  issued_at: Iso8601;
}